fix MLIR ODR violations by compiling MLIR from scratch
It turns out that we really need to build MLIR with
`BUILD_SHARED_LIBS=ON`. If we don't do that, there are weird ODR
violations that appear.

This commit changes the build system to do just that by fetching and
compiling MLIR from scratch. This change also brings about the API
changes brought about by the new version of LLVM.

Also, this commit introduces some initial work on running the JIT.