fix memory corruption
The current approach that uses a shared pointer to handle
ownership between python and C++ is not sufficient and leads to memory
corruption.

This change fixes that by wrapping `Graph` into a class that inherits
from `nb::intrusive_base`. I also removed the `CSEPass` after linalg
which causes wrong results.