rewrite and improve the design of the API
Now there is no need to annotate a function as `axon.jit()`, all
tensor expressions are now traced and computed when the user inspects
the data of a tensor.

Now, I need to rewrite the tests, and make this work. Since it is
currently segfaulting when executing the jitted function.