simplify lowering of jit functions
Before this commit we were lowering tensors as tuple<memref, memref> or
tuple<memref> depending on whether it requires a gradient or not. Now,
we have replaced this by "inlining" the tuple.

So instead of `func.func @graph(tuple<memref, memref>, tuple<memref>)`,
it is now lowered into `func.func @graph(memref, memref, memref)`.

This necessitates some changes and refactors in the codegen, primarily
on how to keep track which argument corresponds to the data or gradient
of a tensor.